requirements.txt
openai>=1.14.0
flask>=3.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
gunicorn>=21.2.0
azure-cosmos>=4.7.0
httpx>=0.28.1
